Move logging configuration out of rotkehlchen.py and in own file
