+++ rotkehlchen/rotkehlchen.py
@@ -3,7 +3,6 @@
 import argparse
 import logging.config
 import os
-import re
 import time
 from pathlib import Path
 from typing import Any, Dict, List, Optional, Tuple, Union
@@ -38,7 +37,12 @@ from rotkehlchen.fval import FVal
 from rotkehlchen.greenlets import GreenletManager
 from rotkehlchen.history import PriceHistorian, TradesHistorian
 from rotkehlchen.inquirer import Inquirer
-from rotkehlchen.logging import DEFAULT_ANONYMIZED_LOGS, LoggingSettings, RotkehlchenLogsAdapter
+from rotkehlchen.logging import (
+    DEFAULT_ANONYMIZED_LOGS,
+    LoggingSettings,
+    RotkehlchenLogsAdapter,
+    configure_logging,
+)
 from rotkehlchen.premium.premium import Premium, PremiumCredentials, premium_create_and_verify
 from rotkehlchen.premium.sync import PremiumSyncManager
 from rotkehlchen.transactions import EthereumAnalyzer
@@ -60,9 +64,6 @@ log = RotkehlchenLogsAdapter(logger)
 MAIN_LOOP_SECS_DELAY = 15
 
 
-PYWSGI_RE = re.compile(r'\[(.*)\] ')
-
-
 class Rotkehlchen():
     def __init__(self, args: argparse.Namespace) -> None:
         """Initialize the Rotkehlchen object
@@ -78,73 +79,7 @@ class Rotkehlchen():
         # authenticate or premium server temporarily offline
         self.premium: Optional[Premium] = None
         self.user_is_logged_in = False
-
-        logfilename = None
-        if args.logtarget == 'file':
-            logfilename = args.logfile
-
-        class PywsgiFilter(logging.Filter):
-            def filter(self, record: logging.LogRecord) -> bool:
-                """Filter out the additional timestamp put in by pywsgi
-
-                This is really a hack to fix https://github.com/rotki/rotki/issues/1192
-
-                It seems that the way they do the logging in pywsgi they create the log
-                entry completely on their own. So the %message part of the entry contains
-                everything and is hence not properly customizale via normal python logging.
-
-                Other options apart from using this filter would be:
-                - Ignore it and just have the timestamp two times in the logs
-                - Completely disable pywsgi logging and perhaps move it all to the
-                rest api.
-                """
-                record.msg = PYWSGI_RE.sub('', record.msg)
-                return True
-
-        loglevel = args.loglevel.upper()
-        formatters = {
-            'default': {
-                'format': '[%(asctime)s] %(name)s: %(message)s',
-                'datefmt': '%d/%m/%Y %H:%M:%S %Z',
-            },
-        }
-        handlers = {
-            'file': {
-                'class': 'logging.FileHandler',
-                'filename': logfilename,
-                'mode': 'w',
-                'level': 'DEBUG',
-                'formatter': 'default',
-            },
-        }
-        filters = {
-            'pywsgi': {
-                '()': PywsgiFilter,
-            },
-        }
-        loggers = {
-            '': {  # root logger
-                'level': loglevel,
-                'handlers': ['file'],
-            },
-            'rotkehlchen.api.server.pywsgi': {
-                'level': loglevel,
-                'handlers': ['file'],
-                'filters': ['pywsgi'],
-            },
-        }
-        logging.config.dictConfig({
-            'version': 1,
-            'disable_existing_loggers': False,
-            'filters': filters,
-            'formatters': formatters,
-            'handlers': handlers,
-            'loggers': loggers,
-        })
-
-        if not args.logfromothermodules:
-            logging.getLogger('urllib3').setLevel(logging.CRITICAL)
-            logging.getLogger('urllib3.connectionpool').setLevel(logging.CRITICAL)
+        configure_logging(args)
 
         self.sleep_secs = args.sleep_secs
         if args.data_dir is None:
